Repositorio/requirements.txt
PyPDF2==3.0.0
python-docx
tk
imaplib
email
regex